Added error codes and memory indices, updated README

+++ disas.c
@@ -1,6 +1,6 @@
 #include "disas.h"
 
-uint16_t *saveBinData(const char *path, uint32_t size) {
+uint16_t *saveBinData(const char *path, uint64_t size) {
 	FILE *bin = fopen(path, "rb");
 	if(bin == NULL) return NULL; // if file cannot be opened, return NULL
 	uint16_t *binData = malloc((size / 2) * sizeof(uint16_t));
